cstomer, item validation complete

+++ src/main/java/com/pahana/persistence/model/Item.java
@@ -1,40 +1,33 @@
 package com.pahana.persistence.model;
 
+import java.math.BigDecimal;
+
 public class Item {
-    private int itemId;
-    private String itemName;
-    private String description;
-    private double costPrice;
-    private double retailPrice;
-    private int quantity;
+    private Integer itemId;         // AUTO_INCREMENT PK
+    private String itemName;        // required
+    private String description;     // optional
+    private BigDecimal costPrice;   // required >= 0
+    private BigDecimal retailPrice; // required >= costPrice
+    private Integer quantity;       // >= 0 (nullable)
 
     public Item() {}
 
-    public Item(int itemId, String itemName, String description, double costPrice, double retailPrice, int quantity) {
-        this.itemId = itemId;
-        this.itemName = itemName;
-        this.description = description;
-        this.costPrice = costPrice;
-        this.retailPrice = retailPrice;
-        this.quantity = quantity;
+    public Item(Integer itemId, String itemName, String description,
+                BigDecimal costPrice, BigDecimal retailPrice, Integer quantity) {
+        this.itemId = itemId; this.itemName = itemName; this.description = description;
+        this.costPrice = costPrice; this.retailPrice = retailPrice; this.quantity = quantity;
     }
 
-    // Getters and Setters
-    public int getItemId() { return itemId; }
-    public void setItemId(int itemId) { this.itemId = itemId; }
-
+    public Integer getItemId() { return itemId; }
+    public void setItemId(Integer itemId) { this.itemId = itemId; }
     public String getItemName() { return itemName; }
     public void setItemName(String itemName) { this.itemName = itemName; }
-
     public String getDescription() { return description; }
     public void setDescription(String description) { this.description = description; }
-
-    public double getCostPrice() { return costPrice; }
-    public void setCostPrice(double costPrice) { this.costPrice = costPrice; }
-
-    public double getRetailPrice() { return retailPrice; }
-    public void setRetailPrice(double retailPrice) { this.retailPrice = retailPrice; }
-
-    public int getQuantity() { return quantity; }
-    public void setQuantity(int quantity) { this.quantity = quantity; }
+    public BigDecimal getCostPrice() { return costPrice; }
+    public void setCostPrice(BigDecimal costPrice) { this.costPrice = costPrice; }
+    public BigDecimal getRetailPrice() { return retailPrice; }
+    public void setRetailPrice(BigDecimal retailPrice) { this.retailPrice = retailPrice; }
+    public Integer getQuantity() { return quantity; }
+    public void setQuantity(Integer quantity) { this.quantity = quantity; }
 }
